% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{predict.bigKRLS}
\alias{predict.bigKRLS}
\title{predict.bigKRLS}
\usage{
\method{predict}{bigKRLS}(object, newdata, se.pred = FALSE, ytest = NULL,
  ...)
}
\arguments{
\item{object}{bigKRLS output}

\item{newdata}{new data. ncol(X) == ncol(newdata) but nrow(X) need not be the same as nrow(newdata).}

\item{se.pred}{get standard errors on predictions?}

\item{ytest}{Provide testing data to have it returned with the object. Optional. To automatically generate out-of-sample test statistics, use crossvalidate.bigKRLS() instead.}

\item{...}{ignore}
}
\value{
Returns bigKRLS_predicted list object.
}
\description{
Predict function for bigKRLS object. crossvalidate.bigKRLS() provides additional functionality.
}
\examples{
 
N <- 500  # proceed with caution above N = 5,000 for system with 8 gigs made avaiable to R
P <- 4
X <- matrix(rnorm(N*P), ncol = P)
b <- 1:P 
y <- sin(X[,1]) + X \%*\% b + rnorm(N)
fit <- bigKRLS(y, X, Ncores = 1)
newdata <- matrix(runif(N * P), ncol = P)
p <- predict(fit, newdata, se.pred = TRUE)
range(p$predicted)
range(p$se.pred)
}
