% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{save.bigKRLS}
\alias{save.bigKRLS}
\title{save.bigKRLS}
\usage{
save.bigKRLS(object, model_subfolder_name, overwrite.existing = FALSE,
  noisy = TRUE)
}
\arguments{
\item{object}{bigKRLS output (regression, prediction, and crossvalidation). Use load.bigKRLS(model_subfolder_name), not load().}

\item{model_subfolder_name}{A name of a folder where the file(s) will be written.}

\item{overwrite.existing}{Logical -- write over folders with the same name? Default == FALSE.}

\item{noisy}{Logical -- display progress, additional instructions? Default == TRUE.}
}
\description{
save function, recommended when bigKRLS output contains big matrices (once N > 2,500 the kernel is stored this way).
Base R data will be stored in a list in an .RData file, big matrices will be stored in .txt files. 
Call load.bigKRLS() to retrieve.
}
\examples{
N <- 500  # proceed with caution above N = 5,000 for system with 8 gigs made avaiable to R
P <- 4
X <- matrix(rnorm(N*P), ncol = P)
b <- runif(ncol(X)) 
y <- X \%*\% b + rnorm(N)
out <- bigKRLS(y, X, Ncores=1)
# save.bigKRLS(out, "exciting_results") # don't use save()
# load.bigKRLS("exciting_results")
}
