% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{shiny.bigKRLS}
\alias{shiny.bigKRLS}
\title{shiny.bigKRLS}
\usage{
shiny.bigKRLS(out, export = F, main.label = "bigKRLS estimates",
  plot.label = NULL, xlabs = NULL, font_size = 20, hline = 0,
  shiny.palette = c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00",
  "#FFFF33", "#A65628", "#F781BF", "#999999"))
}
\arguments{
\item{out}{bigKRLS output. Does not require any N * N matrices.}

\item{export}{Logical -- instead of running Shiny, prepare the key estimates as a separate file? (The N * Ns are too large for most Shiny servers but the key estimates are only N * P).}

\item{main.label}{Main label (upper left of app)}

\item{plot.label}{Optional character}

\item{xlabs}{Optional character vector for the x variables.}

\item{font_size}{Font size. Default == 20. "theme_minimal(base_size = font_size)"}

\item{hline}{horiztonal line. Default == 0 (x axis)}

\item{shiny.palette}{color scheme for main app. 9 colors.
@export}
}
\description{
shiny.bigKRLS
}
