% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{summary.bigKRLS_CV}
\alias{summary.bigKRLS_CV}
\title{summary.bigKRLS_CV}
\usage{
\method{summary}{bigKRLS_CV}(object, ...)
}
\arguments{
\item{object}{bigKRLS_CV output. If you saved with save.bigKRLS(), only the .RData file is needed for this function (for K folds CV, that means only the .RData in the top level folder).}

\item{...}{Additional parameters to be passed to summary() for the training model(s) contained within the CV object. For example, summary(cv, digits = 3).}
}
\description{
Summary function for bigKRLS crossvalidated output.
}
\examples{
N <- 500  # proceed with caution above N = 5,000 for system with 8 gigs made avaiable to R
P <- 5
X <- matrix(rnorm(N*P), ncol=P)
b <- runif(ncol(X)) 
y <- X \%*\% b + rnorm(nrow(X))
cv.out <- crossvalidate.bigKRLS(y, X, seed = 123, ptesting = 10, Ncores = 1)
summary(cv.out)
kcv.out <- crossvalidate.bigKRLS(y, X, seed = 123, Kfolds = 3, Ncores = 1)
summary(kcv.out, labs = c("Alpha", "Beta", "Gamma", "Delta", "Epsilon"))
}
