% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{bigKRLS}
\alias{bigKRLS}
\title{Kernel Regularized Least Squares with Big Matrices}
\usage{
bigKRLS(X = NULL, y = NULL, lambda = NULL, sigma = NULL,
  derivative = TRUE, binary = TRUE, vcov = TRUE, print.level = 3,
  L = NULL, U = NULL, tol = NULL, eigtrunc = NULL)
}
\arguments{
\item{X}{A matrix of observations of the independent variables; missing values not allowed}

\item{y}{A vector of observations on the dependent variable; missing values not allowed}

\item{lambda}{Regularization parameter. Default: estimated based (in part) on the eigenvalues of the kernel}

\item{sigma}{Bandwidth, shorthand for sigma squared. Default: sigma <- ncol(X). Since x variables are standardized, faciltates interprepation of the kernel.}
}
\value{
bigKRLS object containing slope and uncertainty estimates
}
\description{
Kernel Regularized Least Squares with Big Matrices
}
\examples{
N <- 5000  # proceed with caution above N = 5,000
k <- 4
X <- matrix(rnorm(N), ncol=k)
X <- cbind(X, sample(0:1, replace = T, size = nrow(X)))
y <- X \%*\% runif(ncol(X)) + rnorm(N/k)
out <- bigKRLS(X = X, y = y)
}

