% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{load.bigKRLS}
\alias{load.bigKRLS}
\title{load.bigKRLS}
\usage{
load.bigKRLS(path, newname = NULL, pos = 1)
}
\arguments{
\item{path}{Path to folder where bigKRLS object was saved.}

\item{newname}{If NULL (default), bigKRLS object will appear as 'bigKRLS_out'}

\item{pos}{position. Default == 1 (global environment).}
}
\description{
Reconstructs bigKRLS output object as list.
}
