% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{load.bigKRLS}
\alias{load.bigKRLS}
\title{load.bigKRLS}
\usage{
load.bigKRLS(path, newname = NULL, pos = 1, noisy = TRUE,
  return_object = FALSE)
}
\arguments{
\item{path}{Path to folder where bigKRLS object was saved.}

\item{newname}{If NULL (default), bigKRLS regression and prediction output will appear as "bigKRLS_out", while crossvalidation results will appear as "object".}

\item{pos}{position. Default == 1 (global environment). NULL means don't assign (return only).}

\item{noisy}{Logical: display updates?}

\item{return_object}{Logical: return library(bigKRLS) object? Default == FALSE.}
}
\description{
Reconstructs bigKRLS output object as list.
}
\examples{
# save.bigKRLS(out, "exciting_results") # don't use save()
# load.bigKRLS("exciting_results") # don't use load()
}
