% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{predict.bigKRLS}
\alias{predict.bigKRLS}
\title{predict.bigKRLS}
\usage{
\method{predict}{bigKRLS}(object, newdata, se.fit = FALSE, ...)
}
\arguments{
\item{object}{bigKRLS output}

\item{newdata}{new data. ncol(X) == ncol(newdata) but nrow(X) need not be the same as nrow(newdata).}

\item{se.fit}{get standard errors on predictions?}

\item{...}{ignore}
}
\description{
predict.bigKRLS
}
