% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{save.bigKRLS}
\alias{save.bigKRLS}
\title{save.bigKRLS}
\usage{
save.bigKRLS(object, model_subfolder_name, overwrite.existing = F)
}
\arguments{
\item{object}{bigKRLS output}

\item{model_subfolder_name}{A name of a folder where the file(s) will be written.}

\item{overwrite.existing}{Logical -- write over folders with the same name? Default == FALSE.}
}
\description{
save function, recommended when bigKRLS output contains big matrices (once N > 2,500 the kernel is stored this way).
Base R data will be stored in a list in an .rdata file, big matrices will be stored in .txt files. 
Call load.bigKRLS() to retrieve.
}
