% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{save.bigKRLS}
\alias{save.bigKRLS}
\title{save.bigKRLS}
\usage{
save.bigKRLS(object, model_subfolder_name, overwrite.existing = FALSE,
  noisy = TRUE)
}
\arguments{
\item{object}{bigKRLS output (regression, prediction, and crossvalidation). Use load.bigKRLS(model_subfolder_name), not load().}

\item{model_subfolder_name}{A name of a folder where the file(s) will be written.}

\item{overwrite.existing}{Logical -- write over folders with the same name? Default == FALSE.}

\item{noisy}{Logical -- display progress, additional instructions? Default == TRUE.}
}
\description{
save function, recommended when bigKRLS output contains big matrices (once N > 2,500 the kernel is stored this way).
Base R data will be stored in a list in an .RData file, big matrices will be stored in .txt files. 
Call load.bigKRLS() to retrieve.
}
\examples{
# y <- as.matrix(ChickWeight$weight)
# X <- matrix(cbind(ChickWeight$Time, ChickWeight$Diet == 1), ncol = 2)
# out <- bigKRLS(y, X, Ncores=1)
# save.bigKRLS(out, "exciting_results") 
# don't use save() unless out$has.big.matrices == FALSE
# load.bigKRLS("/path/to/exciting_results") 
# path not necessary if in current working directory
}
