% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{summary.bigKRLS}
\alias{summary.bigKRLS}
\title{summary.bigKRLS}
\usage{
\method{summary}{bigKRLS}(object, probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  digits = 4, labs = NULL, ...)
}
\arguments{
\item{object}{bigKRLS output. If you saved with save.bigKRLS(), only the .rdata file is needed for this function.}

\item{probs}{For quantiles.}

\item{digits}{Number of signficant digits.}

\item{labs}{Optional vector of x labels.}

\item{...}{ignore}
}
\description{
Summary function for bigKRLS output. Call knitr::kable(summary(my_ouput)[[1]]) or knitr::kable(summary(my_ouput)[[2]]) to format with RMarkdown.
}
