% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{summary.bigKRLS_CV}
\alias{summary.bigKRLS_CV}
\title{summary.bigKRLS_CV}
\usage{
\method{summary}{bigKRLS_CV}(object, ...)
}
\arguments{
\item{object}{bigKRLS_CV output. If you saved with save.bigKRLS(), only the .RData file is needed for this function (for K folds CV, that means only the .RData in the top level folder).}

\item{...}{Additional parameters to be passed to summary() for the training model(s) contained within the CV object. For example, summary(cv, digits = 3).}
}
\description{
Summary function for bigKRLS crossvalidated output.
}
\examples{
y <- as.matrix(ChickWeight$weight)
X <- matrix(cbind(ChickWeight$Time, ChickWeight$Diet == 1), ncol = 2)
cv.out <- crossvalidate.bigKRLS(y, X, seed = 123, ptesting = 20, Ncores = 1)
cv.out$pseudoR2_oos # cor(y[test], cv.out$tested$predicted)^2 == 0.7488783
# summary(cv.out, labs = c("Alpha", "Beta", "Gamma", "Delta", "Epsilon"))
# cv = summary(cv.out, labs = c("Alpha", "Beta", "Gamma", "Delta", "Epsilon"))
# cv$training.ttests         # hypothesis tests, training model
# kcv.out <- crossvalidate.bigKRLS(y, X, seed = 123, Kfolds = 3, Ncores = 1)
# summary(kcv.out)           # to display
# kcv <- summary(kcv.out, digits = 3) 
# kcv$overview                 # test statistics, in and out of sample all folds
# kcv$training2.ttests         # hypothesis tests, fold 2 
}
