% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigKRLS.R
\name{summary.bigKRLS_CV}
\alias{summary.bigKRLS_CV}
\title{summary.bigKRLS_CV}
\usage{
\method{summary}{bigKRLS_CV}(object, ...)
}
\arguments{
\item{object}{bigKRLS_CV output.}

\item{...}{Additional parameters to be passed to summary() for the training model(s). For example, summary(cv, digits = 3). See ?bigKRLS.summary for details.}
}
\description{
Summary function for bigKRLS crossvalidated output.
}
\examples{
# y <- as.matrix(ChickWeight$weight)
# X <- matrix(cbind(ChickWeight$Time, ChickWeight$Diet == 1), ncol = 2)

# cv.out <- crossvalidate.bigKRLS(y, X, seed = 123, ptesting = 20)
# summary(cv.out)

# cv <- summary(cv.out, labs = c("Alpha", "Beta", "Gamma", "Delta", "Epsilon"))
# cv$training.ttests

# kcv.out <- crossvalidate.bigKRLS(y, X, seed = 123, Kfolds = 3)
# summary(kcv.out)  
        
# kcv <- summary(kcv.out) 
# kcv$overview
# kcv$training2.ttests
}
